Ensure a conn is always returned
When overriding the default `:filter_traces/1` function, specifically when returning false an error occurs.
This is due to the fact that the StartTrace plug `call` function only returns a `conn` for cases when
`filter_traces/1` returns true.

The error happens at the `Plug.Conn.register_before_send/2` function as no `conn` is being passed:

```
** (FunctionClauseError) no function clause matching in Plug.Conn.register_before_send/2
```