Better calculate the route for the resource name
Right now, the calculcation of the routes strips static
parts of the path when calculating the resource name

For instance if you have the following route:

```
scope "/carts" do
  forward "/cart", Routers.CartRouter
end
```

and in the CartRouter
```
scope "/", ApiV1Web do
  get "/:app_id/:identifier", CartController, :get
end
```

the request path would be something like `/carts/cart/1234/abc`. This
would get send with a resource name of `GET /:app_id/:identifier`, which
make is very hard to differentiate between routes in a different
router or scope.

This change preserves the static parts of the route so that a clearer
route name is obtained.