Handle Plug.Parsers.UnsupportedMediaTypeError (#41)
UnsupportedMediaTypeError is raised when Plug.Parsers doesn't support
the content-type specified in the request.

Similarly to Phoenix.Router.NoRouteError, this shouldn't be treated
as an error span, since anyone can send an unsupported media type to
the server. Unlike NoRouteError, we can keep the resource name.

This change matches the UnsupportedMediaTypeError and skips marking the
trace as an error.